get region and hardware options from BE
