+++ packages/toolkit/src/lib/vdp-sdk/model/types.ts
@@ -104,3 +104,8 @@ export type ModelsWatchState = Record<string, ModelWatchState>;
 export type ModelWithPipelines = Model & {
   pipelines: Pipeline[];
 };
+
+export type ModelRegion = {
+  region_name: string;
+  hardware: string[];
+}
