+++ packages/toolkit/src/view/model/ModelHubCreatePageMainView.tsx
@@ -4,7 +4,6 @@ import { useParams } from "next/navigation";
 import { PageTitle } from "../../components";
 import { GeneralAppPageProp } from "../../lib";
 import { CreateModelForm } from "./CreateModelForm";
-import { CreateModelForm as CreateModelForm1 } from "./CreateModelForm1";
 
 export type ModelHubCreatePageMainViewProps = GeneralAppPageProp & {
   disabledCreateModel: boolean;
@@ -25,20 +24,10 @@ export const ModelHubCreatePageMainView = (
           className="mb-5"
         />
       </div>
-      <CreateModelForm1
+      <CreateModelForm
         accessToken={accessToken}
         enabledQuery={enableQuery}
       />
-      {/* <CreateModelForm
-        width="w-full"
-        onCreate={(initStore) => {
-          initStore();
-          router.push(`/${entity}/models`);
-        }}
-        accessToken={accessToken}
-        enabledQuery={enableQuery}
-        disabledCreateModel={disabledCreateModel}
-      /> */}
     </div>
   );
 };
